archive/week2-revert-2026-03-05/crypto-trader/requirements.txt
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
PyJWT>=2.8.0
cryptography>=41.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9
